+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `AsgiBot` is now (by default) started and closed based on the ASGI lifespan events with
+  the `asgi_managed` keyword argument to `AsgiBot.__init__` allowing this to be disabled.
 
 ## [1.0.5a1] - 2021-12-21
 ### Added